fix(sycl): make the default model run on Intel Arc (cambi/speed twins, fp64-free speed extractors) (#1307)
* fix(sycl): eliminate fp64 in speed extractors and fix cambi default model crashes

* test(sycl): add regression test for default model execution under sycl backend

* docs(sycl): record ADR-1179, research digest, and changelog for default model fixes

* docs(sycl): state the measured CPU/SYCL parity instead of bit-exact claims; add the option-table sync invariant

Verifier corrections to the agent's ADR-1179, research digest, docs/backends/sycl/overview.md
and docs/state.md: the default model on SYCL (Arc A380) is identical at %.6f on the 1080p
checkerboard pairs and drifts 2.66e-3 on cambi / 2.0e-3 on vmaf for the 576x324 src01 pair;
the Recently-closed row T-SYCL-V1-MODEL-SEGFAULT-2026-09-04 was missing (only the _Updated
line existed). core/src/feature/sycl/AGENTS.md gains the per-feature option-table sync
invariant (every option the CPU twin declares that a model can set must exist in the SYCL
twin's table, or the model's feature name diverges).

---------

+++ core/src/feature/sycl/speed_chroma_sycl.cpp
@@ -71,11 +71,11 @@ static void launch_means(sycl::queue &q, const float *plane, float *means, uint3
                 return;
             const uint32_t er = elem / SP_BLOCK_SIZE;
             const uint32_t ec = elem % SP_BLOCK_SIZE;
-            double acc = 0.0;
+            float acc = 0.0f;
             for (uint32_t i = 0; i < submatrix_h; ++i)
                 for (uint32_t j = 0; j < submatrix_w; ++j)
-                    acc += (double)plane[(er + i) * stride_px + (ec + j)];
-            means[elem] = (float)(acc / (double)(submatrix_w * submatrix_h));
+                    acc += plane[(er + i) * stride_px + (ec + j)];
+            means[elem] = acc / (float)(submatrix_w * submatrix_h);
         });
     });
 }
@@ -97,7 +97,7 @@ static void launch_cov(sycl::queue &q, const float *plane, const float *means, f
     /* 625 work-groups of COV_WG threads, one per (x_index, y_index) pair. */
     const size_t total_wg = SP_ELEMENTS * SP_ELEMENTS;
     q.submit([&](sycl::handler &cgh) {
-        sycl::local_accessor<double, 1> s_partial(sycl::range<1>(COV_WG), cgh);
+        sycl::local_accessor<float, 1> s_partial(sycl::range<1>(COV_WG), cgh);
         cgh.parallel_for(sycl::nd_range<1>(total_wg * COV_WG, COV_WG), [=](sycl::nd_item<1> it) {
             const uint32_t x_index = (uint32_t)(it.get_group(0) / SP_ELEMENTS);
             const uint32_t y_index = (uint32_t)(it.get_group(0) % SP_ELEMENTS);
@@ -107,16 +107,16 @@ static void launch_cov(sycl::queue &q, const float *plane, const float *means, f
             const uint32_t xc = x_index % SP_BLOCK_SIZE;
             const uint32_t yr = y_index / SP_BLOCK_SIZE;
             const uint32_t yc = y_index % SP_BLOCK_SIZE;
-            const double mean_x = (double)means[x_index];
-            const double mean_y = (double)means[y_index];
+            const float mean_x = means[x_index];
+            const float mean_y = means[y_index];
 
             const uint32_t total = submatrix_h * submatrix_w;
-            double local_sum = 0.0;
+            float local_sum = 0.0f;
             for (uint32_t p = tid; p < total; p += COV_WG) {
                 const uint32_t i = p / submatrix_w;
                 const uint32_t j = p % submatrix_w;
-                const double vx = (double)plane[(xr + i) * stride_px + (xc + j)];
-                const double vy = (double)plane[(yr + i) * stride_px + (yc + j)];
+                const float vx = plane[(xr + i) * stride_px + (xc + j)];
+                const float vy = plane[(yr + i) * stride_px + (yc + j)];
                 local_sum += (vx - mean_x) * (vy - mean_y);
             }
             s_partial[tid] = local_sum;
@@ -128,7 +128,7 @@ static void launch_cov(sycl::queue &q, const float *plane, const float *means, f
                 it.barrier(sycl::access::fence_space::local_space);
             }
             if (tid == 0u)
-                cov_mat[x_index * SP_ELEMENTS + y_index] = (float)(s_partial[0] / (double)total);
+                cov_mat[x_index * SP_ELEMENTS + y_index] = s_partial[0] / (float)total;
         });
     });
 }
